drivers/net/fec: fix RX stall
The FEC sees unmasking RXF events in the EIMR while RDAR is cleared as
an error, and stops issuing RXF interrupts when it detects such
situation. Since the receive path is fully interrupt-driven, this
usually leads to a complete stall.

The combination of high network load on the receiver side, and delays
on the current CPU (kernel preemption, IRQ activity) may cause the RX
ring to fill up again soon after we pulled some/all of the pending
frames, but _before_ RXF is unmasked in the EIMR, leading to the
obnoxious RX stall.

To prevent this, we pull exactly one frame from the RX ring if RDAR is
zero, _after_ the EIMR was updated. RDAR gets written to as a
consequence of such action, which is enough to re-enable the RXF
event.