clocksource: mmio: add user-mappable mmio clocksources
This is a building block for exporting MMIO clocksources to
applications via the vDSO, which we direly need for gaining
syscall-less clock readouts with many types of ARM clocksources.

A new clocksource description (struct clocksource_user_mmio) is
introduced, which specifies how to read the underlying time counter
from a set of generic MMIO operations. In addition, it exports a basic
ioctl-based information & MMIO mapping request, which is available to
applications for gaining (read-only) access to the counter.

This is Gilles's original implementation, to which I added refcounting
to track usage of mappings across fork().

Signed-off-by: Philippe Gerum <rpm@xenomai.org>