dovetail: add core support
This is the basic interface layer for hosting a high-priority resident
co-kernel, aka "dovetailing".

The co-kernel leverages the interrupt pipeline ability to turn common
device IRQs into pseudo-NMIs from the perspective of the in-band code
(CONFIG_IRQ_PIPELINE), to achieve very low latency for the co-kernel
running on the oob stage.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>