printk: irq_pipeline: safe printk with IRQ pipelining
Defer printk output not to wreck the hard interrupt state, cause
unacceptable latency, or risk a deadlock, if either printk() is called
from the oob stage or IRQs are hard disabled on entry (which covers
the case of running over the root stage holding a hard lock).

Deferral is obtained by following the NMI path in vprintk_func() when
any of the above conditions is true. This relies on the ability for
the oob stage to submit work requests to the in-band stage using the
irq_work_queue() interface.