ARM/mm: cache-l2x0: irq_pipeline: allow for disabling write-allocate
HACK ALERT. Introduce the 'l2x0_write_allocate' kernel switch for
controlling the write-allocate policy. Such policy is known to induce
very high latencies: it typically doubles the worst-case figures on
imx6q with cache revisions of the PL310 controller earlier than r3p2.

By default, the default policy is usually determined by the AWCACHE
attribute settings unless the platform code forced it in the auxiliary
control register. In the common case, write allocation is selected.

The patch conditionally changes this setting as follows:

- if the cache controller is a PL210, or a PL310 advertising revision
  r3p2 or later, keep WA enabled unless forced off by
  l2x0_write_allocate=0.

- if the cache controller is a PL220, or a PL310 advertising a
  revision earlier than r3p2, and WA was NOT forced off, issue an
  alert about possibly higher latency figures.

We do know that disabling WA on SMP may be architecturally wrong on
ARM, but this happens to work on i.MX6 hardware with early L2 cache
controllers.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>