irq_work: irq_pipeline: enable work submission from the oob stage
The lockless algorithm of irq_work_queue() is NMI-safe by design,
which means that calling it from the oob stage should be possible,
provided that the code raising the IRQ work interrupt allows it too.

Instead of fixing up multiple implementations of arch_irq_work_raise()
to support this, and with the interrupt pipeline in place, we can use
a pseudo-interrupt from the synthetic domain for scheduling a work on
the local CPU, instead of a real interrupt.

Doing so is cheaper when pipelining IRQs since this entirely bypasses
the hw logic; this also guarantees immediate delivery after the
interrupts are enabled back for the in-band stage regardless of the
architecture, including for those which do not provide for hw-assisted
IRQ_WORK support (which would otherwise postpone the handling until
the next timer tick comes in).

The net result makes irq_work_queue() available for scheduling work
requests in the in-band stage from the oob stage.