timekeeping: irq_pipeline: add NMI-safe ktime_get_real_fast() helper
We need a way for an activity running in the head stage context to
access the wallclock time, which conceptually means to have a NMI-safe
implementation.

Provide a basic implementation substracting a safe copy of the
wallclock-to-monotonic offset from the monotonic time returned by
ktime_get_mono_fast_ns() which is NMI-safe too.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>