ARM: VDSO: add private per-process storage page
The vDSO extension exporting MMIO clocksources to userland needs to
store per-process data in a private storage.

A writable page is mapped into the address space of each emerging
process as part of the vDSO installation. __get_privpage() returns the
address of this area, which is freely usable by the vDSO code for
storing process-private context data.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>