kasan: irq_pipeline: serialize reports from oob contexts
KASAN may detect issues in code running out-of-band, so we have to
serialize the report handler accordingly by turning the regular
spinlock into a hard lock.

This will trigger massive a latency spot when a report is emitted, but
since kernel sanity is at stake when this happens, such spike is not
the most urgent issue.

This also means that the KASAN report lock becomes a raw lock -
therefore cannot sleep with full preemption enabled.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>