evl/thread: explicitly check for rescheduling following evl_suspend_thread()
We want to get rid of automatic rescheduling in evl_suspend_thread()
when the caller is current.

We need this change to later introduce resource-based locking instead
of relying on the ugly global nklock, which is the first step towards
its elimination.

i.e. on the waiter side:

== current way:

xnlock_get_irqsave(&nklock, flags)

if (!condition)
    evl_suspend_thread(curr, T_PEND, ...);

xnlock_put_irqrestore(&nklock, flags)

== future way:

  lock_irqsave(condition->lock);
  if (!condition) {
      add_to_wait_q(curr, condition->queue);
      do {
             evl_suspend_thread()/evl_sleep_on_syn();
      	     unlock_irqrestore(condition->lock);
             evl_schedule();
      	     lock_irqsave(condition->lock);
      } while (!condition);
  }
  unlock_irqrestore(condition->lock);