evl/sched: substitute sched_lock with common preempt count protection
The legacy scheduler lock implementation inherited from Xenomai is
terminally racy and slow, fix this with a stack-based counter like the
in-band kernel does.

At this chance, drop the requirement to apply a round-robin effect to
runnable threads on scheduling parameters change: this is weird POSIX
stuff which implementation was racy too, and makes no sense rt-wise
anyway.