evl/logger: downgrade from element to utility device
There is no upside in sharing loggers between application processes,
since they do not (and will not) deal with file locking. Therefore
loggers are process-local utilities, so we don't need them to be
visible from the file system.

Downgrade the logger element to a plain single device from the
factory. Each open instance of such device is a logging channel.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>