arm64: irq_pipeline: fix interrupt state logic in do_notify_resume()
Once do_notify_resume() has reloaded the the fpsimd state for the
current task, we must not re-enable hard irqs until we are back to
running on el0. Otherwise, some OOB thread might preempt us on the
path to userland, reloading the fpsimd registers with its own state,
causing current to resume with somebody else's fpsimd state.

IOW, do_notify_resume() is the last opportunity for reloading a task's
fpstate if _TIF_FOREIGN_FPSTATE is set before it resumes from el0,
make sure it stays that way including when pipelining interrupts.

At this chance, the way we deal with the interrupt state logic in this
routine was greatly simplified, which unsurprisingly leads to a saner
behavior.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>