dovetail: altsched: account for per-cpu preempt count
When the preemption is maintained in a per-cpu variable instead of a
per-task one (e.g. x86), the current stage level (STAGE_OFFSET) needs
no fix up upon moving a task between the inband and oob stages, as
such tracking does not depend on the current task context anymore.

Instead, we can fully rely on the tracking already done by the
pipeline core via calls to switch_inband() / switch_oob().

At this chance, dovetail_context_switch() is extended in order to deal
with the outgoing context and stage level tracking (only for
architectures with per-task preempt count), which the out-of-band
scheduler does not need to do anymore.