evl/file: provide basic placeholders for building w/o EVL
It is important to have the kernel build properly when EVL is disabled
in Kconfig, and we want to reduce the noise in the mainline code
involved in setting up the out-of-band file context without
systematically resorting to inlined #ifdefery.

Provide a couple of placeholders to enable this for CONFIG_EVL is off.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>