clocksource: mmio: detect stale ucs mapping ownership
After a fork->exit sequence from a parent process owning an ucs
mapping we could have child_vma->private_data(ucs)->mm == stale_mm,
and that mm_struct could be given to another process later on.

If this happens, the ucs mapping indexed on the stale mm might be
matched when scanning the per-ucs hash table, since the hash key is
based on the mm address. Any random process which is given the
recycled mm_struct could therefore be reusing the stale ucs mapping,
eventually causing invalid accesses as a result of calling
clock_gettime().