clocksource/arm_arch_timer: irq_pipeline: handle non-secure PPI override
The proxy tick device needs clock_event_device.irq to match the actual
IRQ channel receiving the timer notifications, so that it can turn on
out-of-band handling when installing (irq_switch_oob()).

When a non-secure PPI is detected for the architected timer in CP15
access mode, we may end up with the per-CPU clock event device ticking
on the non-secure PPI, although the corresponding clock chip
descriptor still designates the secure one as its source, which messes
up interrupt pipelining big time.

Assume that the non-secure PPI event overrides the secure one when
present, fixing up clock_event_device.irq accordingly. This is mostly
a work-around: we may need to reconsider how the tick source is
tracked for switching it to out-of-band mode at some point.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>