x86: irq_pipeline: handle IRQ cleanup events properly
IRQ_MOVE_CLEANUP_VECTOR is not part of the APIC range for which we
have a virtual interrupt domain, but this is actually the first
external IRQ vector which makes it a low-priority event.

Cleanup events are triggered by the CPU the IRQ was moved to when it
receives the first notification of such interrupt, which implicitly
confirms that the migration is complete hardware-wise.  We need to
handle events on this vector specifically in order to hand them to the
interrupt cleanup code.

NOTE: It is currently assumed that smp_irq_move_cleanup_interrupt()
won't cause latency to skyrocket due to a large number of vectors to
clean up in a single pass while holding the hard vector_lock. The good
news is that there is no reason for other CPUs to grab the vector_lock
as part of out-of-band processing. However, the current CPU handling
the cleanup event might run with IRQs hard disabled for too long if
that happens.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>