evl/monitor: add event mask protocol
This protocol operates the monitor value as a set of boolean event
flags forming a 32bit-wide event group. All bits from a group are
initially set to zero.

An event flag group is a lightweight notification mechanism. The
application can send bitmasks to raise individual bits from the group
(i.e. group_value |= bits), or wait for the group to have at least one
bit set for satisfying the request. In the latter case, the group
value is read then cleared atomically, and the collected bits are
returned to the thread heading the wait queue.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>