evl/poll: drop evl_clear_poll_events()
Callers of the out-of-band polling interface are subject to spurious
wakeups by design, which means that they have to handle the case where
a signaled condition goes stale before execution resumes. On the other
hand, we certainly don't want to miss a wake up event in case the
condition is met again in the meantime (i.e. transient
true->false->true event), therefore making provisions for clearing
such event is asking for trouble.

Since the caller has to check for spurious wake ups, we should wake it
up as soon as a false->true transition is observed. Therefore,
evl_clear_poll_events() is deemed useless and even error-prone: drop
it.