evl/factory: inherit clone device ownership
Per-factory device types are introduced to control the ownership of
element devices. Those are created under the /dev/evl hierachy upon
receipt of the EVL_IOC_CLONE request by the clone device of their
parent factory.

With this change, the ownership of a clone device is inherited by all
element devices it instantiates (threads, xbuf, monitors and so
on). This is useful for enabling EVL services for non-privileged
users, which should only require to set the ownership and permissions
of the control and clone devices appropriately.

Caveat: for inherited non-default ownership to stick in presence of
udev/mdev, make sure to define a rule which prevents the default
root.root to be applied for these devices.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>