dovetail: altsched: fix FPU corruption on transition to oob
A task which resumes on the out-of-band stage as a result of calling
dovetail_leave_inband() won't traverse any common scheduling tail such
as dovetail_context_switch() or the inband path to resume in user
mode.

Therefore, we must allow the arch-specific code involved in alternate
scheduling to perform any required fixups when resuming on the oob
stage from dovetail_leave_inband(), such as reloading the FPU context
for the task.