evl/proxy: enable fixed-size write, polling for output to drain
Connecting a proxy with some special file may impose fixed-size writes
to the latter, e.g. an eventfd would require that 64bit values only be
written. These changes enable the application to set a fixed
granularity the proxy must abide by when writing to the target file,
so that such requirement is honored.

In addition:

- oob_write() may block until enough space is available in the
  transfer buffer to complete the sending, unless O_NONBLOCK is set on
  the proxy file.

- POLLOUT|POLLWRNORM can be monitored for waiting for the output to
drain.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>