dovetail: altsched: allow preemption of inband context switch
Most of the time spent in context_switch() involves switching the
memory context (switch_mm_irqs_off) then the register file
(switch_to). Depending on the hardware, such time can amount to tenths
of microseconds, especially when CPU caches are under pressure.

Introduce a short opportunity for preemption in this otherwise atomic
path right after the mm was switched but before switch_to() is called,
so that out-of-band interrupts can be taken, which in turn allows oob
tasks to preempt eventually. The dovetail_context_switch() machinery
already works in a way which guarantees that any in-band context
preempted in a middle of a switch can safely resume from this
location.