genirq: irq_pipeline: add hard_local_irq_sync()
hard_local_irq_sync() is a generic helper backed by arch-specific code
which should synchronize with pending hard irqs by creating a short
window of opportunity for the CPU to be interrupted, before returning
with hard irqs off once done.

We need this for inserting a preemptible spot into the in-band context
switching code.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>