dmaengine: dovetail: define out-of-band DMA requests
Controlling DMA transfers from the out-of-band stage makes it possible
to implement closed-loop control systems running at high frequency,
without involving any in-band operation.

This commit implements the required bits in the generic dmaengine
layer in order for DMA backends to advertise such capability.

DMA_OOB should be set for a DMA backend which can handle out-of-band
transfers, either via the slave or cyclic interface. The behavior of a
transfer descriptor with respect to out-of-band operations can be
controlled by the following flags passed to
dmaengine_prep_dma_cyclic() and/or dmaengine_prep_slave_sg():

- DMA_OOB_INTERRUPT causes the callback attached to the TX descriptor
  to be invoked from the out-of-band stage upon transfer
  completion. The transfer can be either of cyclic or slave type.

- DMA_OOB_PULSE denotes a slave TX descriptor which should be
  triggered manually by calling dma_pulse_oob() for the channel it is
  issued on in order to actually start the transfer in the DMA
  hardware, preventing dmaengine_async_issue_pending() from doing so
  (the latter should still be called to select the transfer descriptor
  though). Typically, a SPI master controller operated from the
  out-of-band stage would use the pulsed mode in order to repeatedly
  trigger the next out-of-band transfer at due time.

The common pattern for preparing an out-of-band transfer driven by an
I/O peripheral applies, e.g.:

   desc = dmaengine_prep_dma_cyclic(chan, ..., DMA_OOB_INTERRUPT);
   desc->callback = oob_callback;
   desc->callback_param = callback_arg;
   ...
   dmaengine_submit(desc);
   dma_async_issue_pending(chan);

Alternatively, an out-of-band cyclic transfer driven by the host at
arbitrary times should be prepared as follows:

   desc = dmaengine_prep_slave_sg(chan, ..., DMA_OOB_INTERRUPT|DMA_OOB_PULSE);
   desc->callback = oob_callback;
   desc->callback_param = callback_arg;
   ...
   dmaengine_submit(desc);
   dma_async_issue_pending(chan);
   ...
   for (;;) {
       ret = dma_pulse_oob(chan);
   }

Running an out-of-band transfer involves the following limitations:

- no other descriptor can be issued on a channel until the ongoing
  out-of-band transfer is terminated.

- an out-of-band descriptor has no precedence over in-band ones until
  it becomes active on the channel as a result of calling
  dma_async_issue_pending().

NOTE: With the exception of dma_pulse_oob() which can be used from any
execution stage including out-of-band, the common dmaengine API must
be called from in-band context as usual, including for descriptor prep
and termination calls.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>