regmap: dovetail: add support for out-of-band locking
Add support for the hard spinlock type, which can serialize callers
which belong to different interrupt stages (i.e. in-band vs
out-of-band). This assumes the protected sections are short, which
was already a common assumption with regular spinlocks.

Out-of-band locking is enabled by setting config.oob_io to true, which
is only considered in fast_io mode.