evl/wait: defer handling of timeout and break conditions
Handling an aborted wait due to a timeout/break condition directly
from the context signaling the wakeup requires the waitqueue lock to
be held around the call to evl_wakeup_thread(), which leads to a
catch-22 situation: we have to ensure that the wait channel
(thread->wchan) is still valid for locking the waitqueue prior to
calling evl_wakeup_thread() as the abort_wait() handler requires, but
that would also require to serialize with evl_wakeup_thread() running
on (a) different core(s) targeting the same thread to make sure
thread->wchan does not vanish.

For instance, we might have a timeout triggered on the current CPU
while the condition we were waiting for is simulateously satisfied on
a remote one. This would end up with an ABBA locking issue for
enforcing this, between the waitqueue lock abort_wait() requires, and
the one protecting the runqueue state evl_wakeup_thread() takes.

To address this, defer such handling until the waiter resumes,
removing itself from the waitqueue upon detecting an aborted wait on
return from evl_schedule(), which lifts the requirement for the
signaling context to do so. Conversely, callers of evl_wake_up() must
now unlink the waiter explicitly instead of relying on the
abort_wait() callback fired by evl_wakeup_thread(). Careful
assumptions and checks in evl_wait_schedule() ensure that we cannot
lose or duplicate wakeup events.

As a result, drop the annoying abort_wait() callback which caused the
catch-22 situation by design.

This is crucial prep work for enabling per-waitqueue locking, getting
rid of the ugly lock in this layer.