evl/lock: do not disable preemption with evl_spinlock
There is no point in expecting the EVL spinlock API to be usable
across calls to rescheduling points, so disabling preemption when
holding such lock is useless. The core should call evl_schedule()
explicitly when it has to do so and leaking a rescheduling opportunity
there is definitely a bug which has to be fixed and not papered over.

With respect to managing the signal vs wakeup issue in event-based
waits, the EVL-based drivers should use the evl_waitqueue* API for
synchronizing on event receipt, which does and helps doing the right
thing.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>