evl/mutex: fix circular dependency issue
Proper locking order is mutex->lock first, thread->lock next.

We are not done with mutex PI, which still locks the entire PI chain
while walking it, which is inherited from Cobalt and badly naive. For
this reason, lockdep still gives us false positive deadlock warnings
with CONFIG_DEBUG_HARD_LOCKS, due to the current requirement to hold
multiple mutex and thread locks when walking this chain.

We should not have to hold more than two locks of each kind at any
point in time, which means that we should be able to resort to simple
_nested locking annotations for lockdep if need be. Work is pending to
make the PI walk preemptible, which will solve lockdep false positives
in the same move.