evl/thread: switch potentially large cpumask to static storage
When a huge number of CPUs is available (e.g. CONFIG_MAXSMP/x86), we
might overflow the stack with cpumask_t variables, for instance with
stack-based thread init descriptors. Since such descriptor only needs
to refer to a constant cpumask, we can maintain a reference to a
global mask instead of embedding a private object there.