lib/vdso: export MMIO clocksources to applications
Make any MMIO-accessed clocksource which can be mapped to user-space
(clocksource_user_mmio) visible to applications via the generic vDSO.

As a result, any registered user-mappable clocksource can be accessed
via the vDSO for syscall-less clock readouts, in addition to the
architected timer (if any).

This code is basically the generalization of a work by Gilles
Chanteperdrix who originally implemented the feature for the ARM
architecture.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>