evl/syscall: honor vDSO fallback calls in oob mode
In some situations, and particularly on legacy/broken hardware, the
kernel might decide to declare the current vDSO-accessible clocksource
as unstable via the clocksource_watchdog mechanism, or we might even
have no vDSO-accessible clocksource available in the first place.

Currently, such issue would cause any call to clock_gettime() via the
vDSO helper to demote the caller to in-band mode, which is pretty lame
and impedes any usage of EVL on such hardware.

This change detects fallback syscalls to clock_gettime() from the
vDSO, and serves requests for timestamps for the CLOCK_MONOTONIC and
CLOCK_REALTIME clocks directly from the out-of-band stage instead of
demoting and propagating the syscall to in-band. Although this does
not guarantee the fastest access to these clocks compared to a
vDSO-based reading, this still allows legacy hardware to be used with
EVL when no clocksource can be accessed via the vDSO.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>