evl/stax: notify oob exclusion via SIGDEBUG_STAGE_LOCKED
Enabling the T_WOSX bit in the thread mode will cause
SIGDEBUG_STAGE_LOCKED to be sent to the caller on return from any
out-of-band sleep caused by a stage exclusion (i.e. in-band holding
the lock).

This is a debug mechanism aimed at detecting unwanted serialization of
oob threads with in-band activity on accessing a resource protected by
a stax.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>