dovetail: add event trigger on return to user mode
In certain situations, we need a way to force a dovetailed task
running in-band to call back the companion core before it returns to
user-space context.

Calling dovetail_request_ucall() for such task will ensure this,
causing the handle_inband_event() handler to receive the
INBAND_TASK_RETUSER event. dovetail_clear_ucall() clears a pending
request for the current task.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>