arm64: irq_pipeline: skip n/a lockdep check in kernel preemption
When irqs are pipelined, the root stage must be unstalled when
branching to preempt_schedule_irq(), although hard irqs must be off.
Skip the lockdep check in the asm trampoline to preempt_schedule_irq()
when interrupts are pipelined. A series of debug checks will be
applied by the latter, including this one.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>