ARM: dovetail: convert to generic oob syscall bit
Define __OOB_SYSCALL_BIT (0x80000000) as the out-of-band syscall
marker usable across all architectures Dovetail supports, which is not
in the range of native syscall numbers.

This removes a longstanding ARM-specific legacy which used to define
oob syscalls as variations of the __ARM_NR_dovetail special system
call.

Since Dovetail only supports the EABI model, there is no point in
keeping this peculiarity which was primarily aimed at coping with the
OABI, which required the syscall number to be encoded in the SWI
instruction, only leaving 24 bits for this purpose (or even down to 8
bits in thumb mode). The EABI conveys the syscall in r7 instead, which
gives us 32 bits for encoding the oob marker.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>