dovetail: add oob stage notifier on trap exit
We currently notify the companion core on entry of any CPU trap which
is taken for a dovetailed task running on the oob stage via the
oob_trap_notify() routine. Such core may also need a way to figure out
whenever the trap handling code is about to return to the caller,
before unwinding the exception frame. Potential use case would be to
switch the current task back to the out-of-band stage in case it had
to switch in-band in order to run the common trap fixup code.

As a result, handle_oob_trap() was renamed
handle_oob_trap_enter(). oob_trap_finalize() is the exit notifier
call, which invokes the handle_oob_trap_exit() handler the core may
override.

NOTE: the overlap with the semantics of the INBAND_TASK_RETUSER event
is marginal, since the latter is fired on the return path to userland
upon request from the core itself. handle_oob_trap_exit() will be
called by Dovetail unconditionally for finalizing traps raised by any
kernel or user context, directly from the trap handling code.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>