KVM: dovetail: make hypervisor coexist with out-of-band scheduling
This finalizes the framework introduced earlier, which is closely
derived from the I-pipe tree commit gcac2d4cd5 maintained at
https://gitlab.denx.de/Xenomai/ipipe-x86.

This framework provides support for maintaining the KVM host state
consistent whenever the in-band stage is preempted by an out-of-band
scheduling request, which entails leaving guest mode at due time.
With this in place, a kernel can safely run KVM-based virtual machines
and out-of-band tasks concurrently.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>