spidev: evl: enable out-of-band mode
This code illustrates the usage of the out-of-band SPI transfer API in
kernel space, interfacing with the EVL core for synchronizing on
transfer completion events. It extends the SPIDEV ioctl() interface
with a few control requests for running out-of-band SPI transfers,
exporting the I/O buffer to applications via mmap() support.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>