evl: factory: add element visibility
Allow users to decide whether a new element should be made public or
private at creation time, depending on a clone operation attribute.

A public element appears as a cdev in the /dev/evl hierarchy, which
means that it is visible to other processes (which was the static
default before this change). On the contrary, a private element is
only known from the process creating it, although it does appear in
the /sysfs hierarchy regardless.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>