genirq: irq_pipeline: fix build with CONFIG_IRQ_PIPELINE off
In order to avoid ugly #ifdeffery cluttering the code uselessly, we
can use code branches checking for running_oob() or running_inband()
which resolves as constants whenever the pipeline is disabled,
compiling out the conditional block in the former case.

We may regularly find irq_post_inband() in such block from
pipeline-aware IRQ handlers which rely on interrupt propagation to
operate, such as oob-capable DMA event handlers. In case
CONFIG_IRQ_PIPELINE is off, irq_post_inband() would not be
resolved. Fix this by adding a dummy placeholder for this case,
ensuring that a build failure is raised whenever the caller was not
compiled out as it should have been.

In addition, we need to pull <linux/irqflags.h> into the placeholder
section so that irqs_disabled() is resolved.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>