evl/thread: add unblock, demote and yield operations
- EVL_THRIOC_UNBLOCK releases the target thread from any ongoing wait
  on a resource (does not apply to forcible suspension).

- EVL_THRIOC_DEMOTE unblocks and demotes the target thread to
  SCHED_WEAK.

- EVL_THRIOC_YIELD performs sched_yield().

In addition, EVL_THRIOC_SET/CLEAR_MODE are no more restricted to the
current thread over oob, which did not make much sense. Calls from
either in-band or oob are now allowed, and may target an arbitrary
thread.

These changes trigger a bump to ABI 22, although we keep on supporting
21.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>