evl: introduce the observable element
Observables enable the observer design pattern, in which any number of
observer threads can be notified of updates to any number of
observable subjects, in a loosely coupled fashion. In the same move,
an EVL thread becomes in and of itself an observable which can be
monitored for events.

As a by-product, provide support for channeling SIGDEBUG notifications
to threads through their own observable on request instead of, or in
addition to issuing SIGDEBUG. The built-in runtime error detection
capabilities the core provides can feed health monitors, based on the
observability of threads.

The core switches to ABI 23 as a result of these changes.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>