evl/proxy: fix handling of POLLIN by the in-band poll handler
Testing the input ring buffer for (non-)emptiness tells us nothing
about the readability of the target file, which may have data
available we did not pull into the buffer yet. We have to relay such
request to the driver actually handling this file in addition to
checking the ring buffer.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>