drivers/evl: latmus: add sirq latency measurement
This mode measures the delay between the moment a synthetic interrupt
is posted from the oob stage and when it is eventually received by its
in-band handler. When measured under significant pressure, this gives
the typical interrupt latency experienced by the in-band kernel due to
local interrupt disabling.

As a result of this change, the current ABI revision is bumped to 25.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>