evl/sched: core: handle invalid CPU error as thread resumes oob
A thread might resume oob from a CPU which is excluded from the set
dedicated to oob scheduling (evl_cpu_affinity), due to a migration
which occurred while in-band (e.g. sched_setaffinity()).

Make sure the thread is allowed to run shortly on the target CPU
nevertheless, so that it can exit cleanly before returning from
evl_switch_oob(). The previous implementation would leave it hanging
in limbo.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>