rcu: irq_pipeline: handle non-maskable pipeline entry context
From the standpoint of the in-band context, an interrupt entry is
logically similar to a NMI when pipelining is enabled, since it may
preempt the in-band stage almost anywhere as IRQs are only virtually
masked most of the time there, including inside (virtually)
interrupt-free sections.

Enable the RCU code to detect a pipeline entry, handling it like a
NMI, so that we may take the RCU read lock there as well.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>