rcu: irq_pipeline: handle non-maskable pipeline entry context
From the standpoint of the in-band context, an interrupt entry is
logically similar to a NMI when pipelining is enabled, since it may
preempt the in-band stage almost anywhere as IRQs are only virtually
masked most of the time there, including inside (virtually)
interrupt-free sections.

Enable the RCU code to detect a pipeline entry, handling it like a
NMI, so that we may take the RCU read lock there as well.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>

rcu: irq_pipeline: enable read lock sections from out-of-band context

Allow an out-of-band caller to start a RCU read locked section by
forcing a RCU entry from NMI. Conversely, leaving such section unwinds
the NMI entry. Like RCU read lock sections, NMI entries can be nested.

In the same move, stop assuming that any out-of-band context means
that RCU is not watching anymore, as a matter of fact the previous
change would contradict this.

The logic is based on the invariant that the pipeline stage cannot
change while in a RCU read locked section.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>