genirq: irq_pipeline: revert hack to ignore oob ticks in in-band time accounting
The hack in steal_account_process_time() is supposed to fix the
following issue:

  <in-band task A>
	   (context switch)
  	   <oob task B>
	   	[timer IRQ, logged for in-band stage]
		...
	   (context switch)
  <in-band task A>
  	   synchronize_irqs(i.e. play events logged for the in-band stage)
		account_process_times(current = task A)

IOW, the un-corrected pattern would cause task A to be charged for all
timer ticks which were actually received by task B while it was running
oob. The code in steal_account_process_time() tries to prevent this by
asking the callers not to charge the CPU time upon timer ticks received
from the oob stage (e.g. X86_IF clear in eflags).

For this reason, maxtime is returned when we don't want the CPU time to
be charged. Problem is that this implementation appeared for kernel
v4.7, and it does not work the way it should with recent kernels anymore
(vtime may be miscalculated for instance).

Drop this code, plan for revisiting the issue in a better way.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>