genirq: irq_pipeline: plug IRQ synchronization lag while idling
default_idle_call() is entered with hard irqs on, in-band stage
stalled, which means that we may have IRQs pending before calling the
arch-specific idling code.

This case is properly detected by irq_cpuidle_enter(), but it fails to
synchronize the in-band log, which causes the pending events to wait
for the next interrupt before they can be played as a result of
unblocking from the CPU idling code, then unstalling the stage, which
is definitely wrong.

irq_cpuidle_enter() must synchronize the in-band log if IRQs are
pending for the stage immediately. In addition, this routine should
always leave the in-band stage unstalled on exit, to align the logic
with the non-pipelined case.

At this chance, also fix a couple of coding style and comment issues.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>