x86: irq_pipeline: defer IRQ affinity update to in-band
Some irqchips require IRQ affinity to be set from the context of the
migrated interrupt, such as Intel's IO*APIC hardware (see
CONFIG_GENERIC_PENDING_IRQ). Since the code actually changing the
affinity may be available to the in-band stage exclusively, we have to
defer its execution until events are synchronized for that stage
before returning from the interrupt frame.

The IO-APIC ack handlers now schedule an irq_work routine to perform
any pending affinity update which is guaranteed to run on top of the
current interrupt frame, when handle_irq_pipelined_finish() eventually
synchronizes the in-band stage before leaving. This makes possible to
change the affinity of an interrupt otherwise delivered to the oob
stage.

It might happen that such update is skipped in case the context
preempted by the interrupt should be hidden from the in-band stage,
such as running oob or when the stage is stalled. If so, the update
would wait for the next interrupt to happen in the right context, like
the idle loop or any other irq-preemptible section of the kernel code.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>