genirq: irq_pipeline: keep in-band stalled until arch_cpu_idle()
We need the whole CPU idling path to run with the in-band stage
stalled until the arch-specific code is invoked to enter the idle
mode.

This fixes a regression introduced by #40e6704, which left
irq_cpuidle_enter() with the in-band stage unstalled. Instead, the
latter should stall the in-band stage before returning, since debug
assertions may expect this later on.

It should be noted that since we ran with hard irqs off, the
regression did not cause any issue with the mere logic. This was more
of a problem with debug assertions checking for irqs_disabled().

Signed-off-by: Philippe Gerum <rpm@xenomai.org>