evl/thread: align trap_entry/exit on new Dovetail rules
The trap_entry handler may expect hard irqs off on entry. It may
demote the current context to the in-band stage, may return with hard
irqs on.

The trap_exit handler only runs if trap_entry was called for the same
trap occurrence. It may expect hard irqs off on entry, may switch the
current context back to the oob stage. Must return with hard irqs
off. This departs from the previous implementation which allowed the
trap_exit handler to return with hard irqs on.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>