genirq: irq_pipeline: fix interrupt disabling logic
The basic idea of altering the IRQ log of remote CPUs underlying
irq_pipeline_clear() is fundamentally flawed. This log contains
strictly per-CPU information, and no other CPU core but the owner
should ever access it. With irq_pipeline_clear() using non-atomic
bitwise ops, failing to abide by this golden rule can only lead to
races causing CPUs to miss logged interrupts and/or play the wrong
ones.

This means that we cannot rely on clearing bits into remote IRQ logs
for preventing disabled IRQs which might already pending to be
played. To fix this issue properly, we observe that:

- the pipeline fully serializes access to the descriptor data on
  desc->lock which we turn into a hard spinlock, so the
  IRQD_IRQ_DISABLED logic is already compatible with any context,
  including the out-of-band stage and pipeline entry code.

- IRQD_IRQ_DISABLED is already used to filter out interrupt delivery
  to handlers from the in-band stage.

Therefore, we only need to extend the logic to the out-of-band
context, checking IRQD_IRQ_DISABLED on this delivery path too,
skipping the handler if set. This also allows disable_irq() to
synchronize cleanly with both in-band and out-of-band events in the
same move.

This bus was detected while porting Dovetail to the octa-core
Exynos5422 processor, where IRQs received from the thermal management
unit (TMU) would be disabled from their handler by a call to
disable_irq_nosync(), eventually causing havoc in the interrupt log of
remote CPU(s), spuriously clearing random interrupt events pending
there, due to non-atomic accesses to its bitmaps.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>