evl: add net core (EXPERIMENTAL)
This code lays the groundwork for an out-of-band network stack.  It
currently provides basic AF_PACKET support (SOCK_RAW only) from the
out-of-band stage, reusing the regular NIC drivers unmodified for
dealing with the hardware. Therefore, the stack has NO end-to-end
real-time property just yet.

The next steps will address:

- a kernel API for making regular NIC drivers oob-capable.

- a mechanism for the out-of-band core to learn the routing
  information dynamically produced by the in-band stack.

- support for UDP/AF_INET.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>