evl/crossing: add oob vs in-band synchronization barrier
This is a simple synchronization mechanism allowing an in-band caller
to pass a point in the code making sure that no out-of-band operations
which might traverse the same crossing are in flight.

Out-of-band callers delimit the danger zone by down-ing and up-ing the
barrier at the crossing, the in-band code should ask for passing the
crossing.

CAUTION: the caller must guarantee that evl_down_crossing() cannot be
invoked _after_ evl_pass_crossing() is entered for a given crossing.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>