evl/observable: convert observable, subscriber locks to hard locks
The subscriber lock is shared between both execution stages, but
accessed from the in-band stage for the most part, which implies
disabling hard irqs while holding it. Meanwhile, out-of-band IRQs and
EVL threads may compete for the observable lock, which would require
hard irqs to be disabled while holding it.  Therefore we would not
generally benefit from the preemption disabling feature we are going
to add to the EVL-specific spinlock in any case. Make these hard locks
to clarify the intent.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>