evl/thread: convert thread lock to hard lock
Out-of-band IRQs and EVL thread contexts may compete for such lock,
which would require hard irqs to be disabled while holding it.
Therefore we would not benefit from the preemption disabling feature
we are going to add to the EVL-specific spinlock. Make it a hard lock
to clarify the intent.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>