ARM: dovetail: enable I-pipe compat for syscall routing
Allow legacy applications to issue syscalls following the I-pipe
syscall convention in EABI mode, which specifies that r7 should be
loaded with 0xf0042 in order to mark out-of-band syscalls.

When CONFIG_IPIPE_COMPAT is enabled, OOB_SYSCALL_BIT is ORed into r7
on the fly if its original value is 0xf0042, so that the syscall will
be routed to the companion core as expected.

This compat mode may be removed on short notice. Do NOT rely on it for
new applications.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>