genirq: irq_pipeline: synchronize log on irq exit to kernel
We must make sure to play any IRQ which might be pending in the
in-band log before leaving an interrupt frame for a preempted kernel
context.

This completes "irq_pipeline: Account for stage migration across
faults", so that we synchronize the log once the in-band stage is
unstalled. In addition, we also care to do this before
preempt_schedule_irq() runs, so that we won't miss any rescheduling
request which might have been triggered by some IRQ we just played.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Suggested-by: Jan Kiszka <jan.kiszka@siemens.com>