evl/factory: define access check for factories
Some operations on factories started from /sys may be sensitive. Add a
predicate to check whether the caller is authorized to perform such
type of operation on a given factory.

Access is granted for CAP_SYS_ADMIN-capable users, and/or users whose
euid matches the kuid of the factory's clone device.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>