dovetail: add prctl() call form
The syscall form used for issuing requests to the companion core is
based on ORing __OOB_SYSCALL_BIT to the syscall register for tagging
an out-of-band syscall, which is outside of the valid in-band range
for syscall numbers.

Unfortunately, this method has a major drawback: tools like Valgrind
which track system calls issued by applications may choke on
out-of-band system calls since these do not belong to the common set
it recognizes. Although adding syscall wrappers to Valgrind might be a
solution, this would not be the most practical one for dual kernel
syscall interfaces. The most practical approach would be to piggyback
on an existing system call which Valgrind (and others) already
recognizes.

To this end, we provide another option for issuing out-of-band
syscalls to the companion core, based on the prctl(2) call. When
caught on kernel entry, this request is unconditionally routed to the
core for inspection, along with any syscall which bears the
__OOB_SYSCALL_BIT tag as previously.

According to this scheme, the core should figure out whether it should
handle the current prctl() request, or leave it to the in-band kernel
(i.e. propagation status set to zero). To fit in the prctl()
signature, the core would typically use the prctl() @option argument
to encode its own syscall number, which would leave enough room for up
to 4 arguments to the corresponding request.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>