dovetail: core: fix COMPAT syscall detection
A syscall number may be different for the same service in native or
compat mode, e.g. aarch64 vs ARM32 for __NR_prctl. As a result, we
cannot check for __NR_prctl verbatim in order to detect out-of-band
syscalls.

Delegate such detection to the architecture layer instead, which
should check for the current task mode _and_ the syscall number for
__NR_prctl from the appropriate native/compat set.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>