irqchip/ti-sci-inta: irq_pipeline: enable pipelined interrupt control
The only actual change is to switch the parent irq for the aggregator
to oob mode, so this can be propagated to the chained irqs if need be
(i.e. IRQF_OOB for them). If a chained irq is not requested by the oob
stage, the oob mode on the parent simply yields no effect.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>