evl/mutex: fix race in release path
The issue is with the mutex owner handle transitioning to
EVL_NO_HANDLE before transfer_ownership() may switch it to the handle
of the first contender in line. If evl_trylock_mutex() or a user
context manage to slip in with a succcessful cmpxchg() in this tiny
window, two threads may actually end up holding the same lock.

To address this bug, set the handle directly to its final value as a
result of transferring ownership, without transitioning via
EVL_NO_HANDLE, closing the race window.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>
Reported-by: Giulio Moro <giulio@bela.io>