evl/mutex: use two-phase tracking on release
Releasing a mutex may involve transferring ownership to the next
waiter in line. This entails the same actions than performed by
locking it, i.e. adding the PP boost if need be and propagating this
change to the wait channel the (new) owner pends on, which are two
separate operations, running under distinct locking conditions.

To this end, update transfer_ownership() and __evl_unlock_mutex() to
use set_mutex_owner() like evl_lock_mutex_timeout() does, with a split
process.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>