evl/mutex: fix mutex counting for T_WOLI|T_WEAK
When T_WOLI/T_WEAK is set for a thread acquiring or releasing a mutex,
the count of mutexes it holds is tracked. Problem is that in some
cases, a thread may release a mutex while bearing these bits, although
it did not have them set when acquiring it, leading to imbalance in
counting and general havoc due to any decision based on that
information afterwards.

We fix this by marking every mutex which participates to this counting
with the new EVL_MUTEX_COUNTED flag, to keep the accounting accurate.

Among several issues, this fixes this kernel splat observed on armv7,
caused by evl_drop_current_ownership() being denied to unlock a mutex
because of a bad tracking count, which in turn would cause its caller
to loop indefinitely:

[   52.576621] WARNING: CPU: 1 PID: 249 at kernel/evl/mutex.c:1383 evl_drop_current_ownership+0x50/0x7c
[   52.585878] Modules linked in:
[   52.589006] CPU: 1 PID: 249 Comm: sched-quota-acc Not tainted 5.15.64-00687-g07ee2d347ee9-dirty #572
[   52.598170] Hardware name: Freescale i.MX6 Quad/DualLite (Device Tree)
[   52.604718] IRQ stage: Linux
[   52.607625] [<c0111c28>] (unwind_backtrace) from [<c010bc44>] (show_stack+0x10/0x14)
[   52.615411] [<c010bc44>] (show_stack) from [<c0ebaedc>] (dump_stack_lvl+0xac/0xfc)
[   52.623017] [<c0ebaedc>] (dump_stack_lvl) from [<c01271e4>] (__warn+0xd4/0x154)
[   52.630357] [<c01271e4>] (__warn) from [<c0eb5018>] (warn_slowpath_fmt+0x60/0xbc)
[   52.637877] [<c0eb5018>] (warn_slowpath_fmt) from [<c02466b4>] (evl_drop_current_ownership+0x50/0x7c)
[   52.647130] [<c02466b4>] (evl_drop_current_ownership) from [<c024dc50>] (cleanup_current_thread+0x60/0x3f4)
[   52.656903] [<c024dc50>] (cleanup_current_thread) from [<c024e83c>] (put_current_thread+0x24/0xc8)
[   52.665891] [<c024e83c>] (put_current_thread) from [<c025241c>] (thread_ioctl+0xa4/0xd0)
[   52.674011] [<c025241c>] (thread_ioctl) from [<c0316bc8>] (sys_ioctl+0x5a8/0xef4)
[   52.681531] [<c0316bc8>] (sys_ioctl) from [<c0100080>] (ret_fast_syscall+0x0/0x1c)

Signed-off-by: Philippe Gerum <rpm@xenomai.org>