evl/mutex: fix unlock vs lock race
In a locking attempt, walking the PI chain is preemptible by other
CPUs since mutex->wchan.lock may be released during the process, so
the latest owner of the mutex might have dropped it while we were busy
progressing down this chain.

Do not block on the mutex if it has no owner on exit from the PI chain
walk, drop of from the wait list then retry a locking attempt from the
start instead.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>