evl/mutex: switch to sleep-and-retry locking pattern
Instead of having evl_unlock_mutex() eagerly transfer the ownership of
a mutex to the next waiter in line directly, wake the latter up and
let it retry the acquisition until it succeeds or an unrecoverable
error happens. This sleep-and-retry pattern has key upsides:

- it does not create any race with a preemptible PI walk running
concurrently, with respect to the current owner of a mutex,
simplifying and ironing the logic there.

- it removes the requirement for handling specifically the mutex steal
case (T_ROBBED+T_WAKEN), where a high priority thread would be allowed
to snatch the ownership until the current owner actually resumes
execution after unblock. This property is naturally enforced by the
new locking pattern, there is no need for specific code since
ownership is always grabbed by the acquiring thread.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>