x86: dovetail: allow for conditional trap handling
Some companion cores may want to fully handle some of the traps
directly from the out-of-band stage, without demoting the current
context to the in-band stage. Typically, #UD, #NM or #DE could be
handled that way. Instead of applying the in-band fixups
unconditionally, assume the following:

- if the companion core switches in-band upon notification, then the
in-band kernel code may assume that all regular fixups need to be
applied next. This is the common case for most traps.

- if the companion core stays on the out-of-band stage upon
notification, then the in-band kernel code should assume that all
required fixups were done from that stage already, returning asap from
the outer trap handler.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>