x86: dovetail: handle WARN directly from the out-of-band stage
Dovetail normally notifies the companion core about most traps
including when WARN[_ON]() has tripped, expecting a transition to the
in-band stage if the CPU was running oob on entry.

This is a problem when it comes to handling debug assertions, since
the core might place them in sections of code where a stage switch
would be unsafe, e.g. because they would hold some hard spinlocks
involved in context switching, leading to a chicken-and-egg situation,
eventually to a deadlock due to a recursive locking attempt.

Let's make handle_bug() oob-safe, skipping the notification when
handling a WARN trap.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>