evl/mutex: account for unbalanced fast_lock/slow_unlock pattern
A thread might have acquired a lock from user-space via the fast
procedure, releasing it from kernel space - this pattern typically
happens with the event monitor, e.g. fast_lock(x) -> signal(event) ->
slow_unlock(x).

In this case, the owner field into the wait channel might be unset if
the lock was uncontended, although the atomic handle does match the
thread's fundle. This is a legit case: skip the PI/PP deboosting if so
because no boost can be in effect for such lock, waking up the heading
waiter if any.

CAUTION: this issue can be noticed only if T_WOLI is disabled for the
thread (which means that CONFIG_EVL_DEBUG_WOLI is off too), otherwise
only the slow acquire/release path would be taken, papering over the
bug.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>