genirq: irq_pipeline: drop IRQ descriptor cache
With the introduction of a maple tree to index IRQ descriptors when
sparse mode is enabled, we may expect cache-friendly operations, at
least better than what was achieved using a radix tree. Some figures
on low-end hardware equiped with not-that-snappy cache subsystems and
running EVL confirmed this.

Let's drop the ad hoc IRQ descriptor cache which is now useless.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>