genirq: irq_pipeline: enable IRQ event forwarding
Adding oob support to an in-band device driver may involve splitting
the existing interrupt handling code between what needs to be done
from the oob stage, and what should wait until the in-band stage is
active. IOW, this is two-step subsequent handling of an IRQ event over
both stages by the same action handler(s). A typical example is DMA
handling, where some channels would require oob handling of related
IRQ events, while others could wait to serve in-band clients instead.

This pattern used to be implemented by a call to irq_post_inband()
from the oob context handler, which is wrong for real devices since
the flow handler would then reissue the hardware acknowledge from the
in-band stage.

To address this, we introduce the IRQ_FORWARD status, which may be
returned by an oob action handler to tell the pipeline that it wants
to be called again for the same event, from the in-band stage.

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>